fix: print smaller types

+++ input/01_var_dec_basic.rs
@@ -1,5 +1,5 @@
 fn main(){
-    let x: i32;
+    let x: i16;
     x = 10;
     println!("{}", x);
     let ab123: i32 = 20;
